Add the home page and new side menu, modify dashboard style
This commit adds a new home page and a new side menu to the application. The side menu has been redesigned with a more modern look and feel. The dashboard has also been modified with new styles and colors to match the new design. The login page still needs to be updated, but this will be done in a future commit. All changes are adjustable and can be modified as needed.